Change ordering of processed
Currently an issue exists within the HA implementation where if multiple
agents are started concurrently, they don't have time to set the object
their downloading to "Processed". By changing the ordering of when we
set an object to processed - before it's downloaded, other threads will
be able to see that the object is being worked on or is already
published.

The caveat here is that if an object fails to publish or download
correctly, it will already be set to "processed". The trade off is
either publishing duplicates, or losing a log or two.