compute wallclock duration for ALB events (#68)
This changes how duration_ms is computed for ALB logs. According to the AWS docs, request_creation_time is the time when the load balancer first gets the request. timestamp is poorly defined, but appears to correspond to when the event is actually finished. Therefor, comparing the two should allow us to see the actual transmission duration as observed by the client. I confirmed that a very long (multi-minute) download I performed does in fact have a multi-minute duration using this calculation.

Also add a test for ALB logs, and add parenthesis to a couple of fields which didn't have them.